Fix datamanager 400 error and equitypricemodel 500 error
1. Datamanager 400 Error on `/predictions` endpoint

Root Cause: The `PredictionQuery` struct in `storage.rs` expected
`timestamp: DateTime<Utc>`, but the portfoliomanager sends timestamps
as Unix floats (e.g., `1737331200.0`). Serde couldn't deserialize the
float into a DateTime.

Fix: Changed `applications/datamanager/src/storage.rs`:
- `PredictionQuery.timestamp` from `DateTime<Utc>` to `f64`
- Added conversion: `DateTime::from_timestamp(prediction_query.timestamp as i64, 0)`

2. Equitypricemodel 500 Error

Root Cause: Tinygrad uses SQLite for disk caching compiled kernels.
FastAPI/uvicorn runs requests in a thread pool, causing:
`sqlite3.ProgrammingError: SQLite objects created in a thread can only
be used in that same thread.`

Fix: Added `DISKCACHE=0` environment variable to:
- `infrastructure/__main__.py` (ECS task definition)
- `infrastructure/docker-compose.yaml` (local development)

This disables tinygrad's disk cache, avoiding the SQLite threading issue.

+++ applications/datamanager/src/data.rs
@@ -8,12 +8,12 @@ use tracing::{debug, info, warn};
 pub struct EquityBar {
     pub ticker: String,
     pub timestamp: i64,
-    pub open_price: Option<u64>,
-    pub high_price: Option<u64>,
-    pub low_price: Option<u64>,
-    pub close_price: Option<u64>,
-    pub volume: Option<u64>,
-    pub volume_weighted_average_price: Option<u64>,
+    pub open_price: Option<f64>,
+    pub high_price: Option<f64>,
+    pub low_price: Option<f64>,
+    pub close_price: Option<f64>,
+    pub volume: Option<f64>,
+    pub volume_weighted_average_price: Option<f64>,
     pub transactions: Option<u64>,
 }
 
@@ -123,7 +123,7 @@ pub fn create_predictions_dataframe(prediction_rows: Vec<Prediction>) -> Result<
 #[derive(Debug, Deserialize)]
 pub struct Portfolio {
     pub ticker: String,
-    pub timestamp: i64,
+    pub timestamp: f64,
     pub side: String,
     pub dollar_amount: f64,
     pub action: String,
@@ -137,7 +137,7 @@ pub fn create_portfolio_dataframe(portfolio_rows: Vec<Portfolio>) -> Result<Data
 
     let portfolio_dataframe = df!(
         "ticker" => portfolio_rows.iter().map(|p| p.ticker.as_str()).collect::<Vec<&str>>(),
-        "timestamp" => portfolio_rows.iter().map(|p| p.timestamp).collect::<Vec<i64>>(),
+        "timestamp" => portfolio_rows.iter().map(|p| p.timestamp).collect::<Vec<f64>>(),
         "side" => portfolio_rows.iter().map(|p| p.side.as_str()).collect::<Vec<&str>>(),
         "dollar_amount" => portfolio_rows.iter().map(|p| p.dollar_amount).collect::<Vec<f64>>(),
         "action" => portfolio_rows.iter().map(|p| p.action.as_str()).collect::<Vec<&str>>(),
